requirements.txt
tqdm
numpy
hdbscan
matplotlib
umap-learn
requests
sentence-transformers
# transformers

# openpyxl
# pandas
# scikit-learn
# pydot
# networkx
# graphviz
